Allow deploy-to-testpypi to fail (#17)
The deployment to testpypi may fail if we try to upload the same version
twice. This scenario can occur if we have to make changes to develop
post merging a feature, but before merging it to master. This error prevents
us from merging our changes into master as the ci job considers it a failure.
Hence the change to the ci behavior to not treat deploy-to-testpypi as a failure.

Co-authored-by: Aaron Alphonso <alphonsoaaron19+git@gmail.com>